OAuth CLI - minor refactoring + fixes
